# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Employee Management System API
  description: |
    RESTful API for managing employees, users, leave requests, and authentication.
    
    Features:
    - Employee CRUD operations with CSV import/export
    - User management with role-based access control
    - Leave request workflow with approval/rejection
    - Email verification and password reset
    - Rate limiting and caching
  version: 1.0.0
  contact:
    name: EMS Development Team
  license:
    name: MIT License
servers:
  - url: https://employee-management-system-production-2a1a.up.railway.app/api/v1
    description: Production server (Railway)
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - cookieAuth: []
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Register a new user with an existing employee ID
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterEmployeeRequest'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate a user and create a session
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Employee Management
...
  /email-verification/send:
    post:
      tags:
        - Email Verification
      summary: Send verification email
      operationId: sendVerificationEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
      responses:
        '200':
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /email-verification/verify:
    post:
      tags:
        - Email Verification
      summary: Verify email
      operationId: verifyEmail
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session-based authentication
  schemas:
...
    RegisterEmployeeRequest:
      type: object
      required:
        - username
        - password
        - employeeId
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "john.doe"
        password:
          type: string
          minLength: 6
          example: "change_me"
        employeeId:
          type: integer
          example: 1
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "admin"
        email:
          type: string
          format: email
          example: "admin@example.com"
        password:
          type: string
          example: "change_me"
    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: "Login successful"
        success:
          type: boolean
          example: true
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        validationErrors:
          type: object
          additionalProperties:
            type: string
...