# src/main/java/com/example/employee_management_system/dto/RegisterEmployeeRequest.java
package com.example.employee_management_system.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;

public class RegisterEmployeeRequest {
    
    @NotBlank(message = "Username is required")
    @Size(min = 3, max = 50, message = "Username must be between 3 and 50 characters")
    private String username;
    
    @NotBlank(message = "Password is required")
    @Size(min = 6, message = "Password must be at least 6 characters long")
    private String password;

    @NotNull(message = "Employee ID is required")
    private Integer employeeId;

    // Default constructor
    public RegisterEmployeeRequest() {}

    // Constructor
    public RegisterEmployeeRequest(String username, String password, Integer employeeId) {
        this.username = username;
        this.password = password;
        this.employeeId = employeeId;
    }

    // Getters and setters
    public String getUsername() { return username; }
    public void setUsername(String username) { this.username = username; }
    
    public String getPassword() { return password; }
    public void setPassword(String password) { this.password = password; }
    
    public Integer getEmployeeId() { return employeeId; }
    public void setEmployeeId(Integer employeeId) { this.employeeId = employeeId; }
}